Convert build to ExtUtils::MakeMaker
Although distzilla is powerful, the files in the actual
distribution aren't saved in the repo, but generated.  Because of
the length of time since the last release, the generated files,
specifically the tests, were quite different for this rebuild.

Since distzilla generates an ExtUtils::MakeMaker file, this commit
converts the build to using it directly.  The generated tests have
been added to the repo and have also been updated to remove a
couple modules as required dependencies.

This commit also adds the other generated files from the dist into
the repo, like LICENSE, META*, README, and fixes the Changes file
to display correctly on metacpan.